% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{are.params.created}
\alias{are.params.created}
\title{check the some variables considered as parameters are created on a server}
\usage{
are.params.created(param_names = c())
}
\description{
check the some variables considered as parameters are created on a server
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
