% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCoordinatesDS.R
\name{assignCoordinatesDS}
\alias{assignCoordinatesDS}
\title{assign data to one or multiple servers with some encrypted data from the analysis computer INCORRECT}
\usage{
assignCoordinatesDS(
  header = "",
  payload = "",
  property.a = 0,
  property.b = 0,
  property.c = 0,
  property.d = 0
)
}
\arguments{
\item{header}{character argument. Header information received from another server.}

\item{payload}{character argument. Payload information received from another server.}

\item{property.a}{numeric argument. Property.a received from another server.}

\item{property.b}{numeric argument. Property.b received from another server.}

\item{property.c}{numeric argument. Property.c received from another server.}

\item{property.d}{numeric argument. Property.d received from another server.}
}
\description{
This server function assigns some values into a specific structure. INCORRECT
}
\details{
Some data are being assign into a specific structure used to share parameter in some privacy-protection settings. The process used by
\link[dsShareServer]{getDataDS} is reversed.
}
\seealso{
\link[dsShareServer]{getDataDS}
}
