% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignParamSettingsDS.R
\name{assignParamSettingsDS}
\alias{assignParamSettingsDS}
\title{assigns some settings used to encrypt and decrypt the parameters}
\usage{
assignParamSettingsDS(param_names = "")
}
\arguments{
\item{param_names}{character vector. Name of the server parameter to encrypt.}
}
\description{
This server function sets some settings specific to the parameters encryption and decryption mechanisms.
The latter should identify a column and row for each parameter in some matrices. The row and column is disclosive. So, it remains
on the server and cannot be analysed directly.
}
