% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{encode.data.no.sharing}
\alias{encode.data.no.sharing}
\title{encode some randomised data}
\usage{
encode.data.no.sharing()
}
\description{
This server-side function generates some random data to be made available to a client-side
function. Its purpose is to mimick the same behaviour as [dsServerParameter]{encode.data.with.sharing}. It aims a
a "decoy", if an error has occurred in the process.
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
\seealso{
\link[dsShareServer]{getDataDS} \link[dsShareServer]{getCoordinatesDS},[dsShareServer]{encode.data.no.sharing}
}
