% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{encode.data.with.sharing}
\alias{encode.data.with.sharing}
\title{encode some obscured  data to be exchanged from one server to another.}
\usage{
encode.data.with.sharing(encrypted.data, length, index)
}
\arguments{
\item{encrypted.data}{- data to be encoded}

\item{length}{- an indication how long should be the data}

\item{index}{- A random number related to the data}
}
\description{
This server function can only be used with some encrypted data. It
format the data prior its transfer to a client-side function.
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
\seealso{
\link[dsShareServer]{getDataDS}, \link[dsShareServer]{getCoordinatesDS},
\link[dsShareServer]{encode.data.no.sharing}
}
