% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptParamDS.R
\name{encryptParamDS}
\alias{encryptParamDS}
\title{encrypt a server parameter}
\usage{
encryptParamDS()
}
\description{
This server function encrypts a given parameter using a dot product and two shared secrets.
}
