% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.transfer}
\alias{get.transfer}
\title{returns the transfer list if it is correctly setup}
\usage{
get.transfer(settings)
}
\arguments{
\item{settings}{some settings}
}
\description{
returns the transfer list if it is correctly setup
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
\note{
Throws error "SERVER::ERR:SHARE::013" if transfer is not created on server.
"Throws "SERVER::ERR:SHARE::014" if the transfer list is created on a server, with the incorrect field.
}
