% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{is.sharing.allowed}
\alias{is.sharing.allowed}
\title{verifies the variables used to set the parametrisation to for sharing parameters
exists on a DataSHIELD server.}
\usage{
is.sharing.allowed()
}
\description{
This server function checks some settings used exchange parameters between
DataSHIELD server exists. It also verifies the data owners and governance have allowed the
sharing of parameters in a specific server.
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
