% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDataEncodedDS.R
\name{isDataEncodedDS}
\alias{isDataEncodedDS}
\title{check some R objects are suitably encoded}
\usage{
isDataEncodedDS(
  data.server = NULL,
  data.encoded = NULL,
  data.held.in.server = NULL
)
}
\arguments{
\item{data.server}{- character argument representing the name of a data frame with the original data}

\item{data.encoded}{- character argument representing the name of a data frame with the encoded data}

\item{data.held.in.server}{- character argument representing the datasets held in a DataSHIELD server as a data frame}
}
\value{
TRUE if the encoding is suitable. Otherwise false.
}
\description{
check some R objects are suitably encoded
}
\details{
This server function verifies the following rules are applied to the encoded data
against (1) a server variable and (2) a datasets held in the server itself.

1. No object is identical
2. None of values the server variable is present in the encoded values
3. All the values are numeric
4. The server data and the encoded data are both significantly different (T-test and
Mann-Whitney Test, p = 0.99)
5. None of the values are within the limit set by the non-disclosure option dsShareServer.near.equal.limit

Additionally the encoded data must hold this condition against the server variable:

6. The encoded data has a greater number of columns than the server variable
}
