% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEndOfDataDS.R
\name{isEndOfDataDS}
\alias{isEndOfDataDS}
\title{Verifies the end of some encoded data on the server has been reached}
\usage{
isEndOfDataDS(data_encoded = NULL)
}
\arguments{
\item{data_encoded}{- character argument specifying the name of the encoded data on a data server}
}
\value{
A boolean value. TRUE if the last row of the encoded data has been transferred.
Otherwise, returns FALSE
}
\description{
This server function indicates whether the number of rows transferred has exceeded
of equal to the number of rows in the encoded data frame.
}
\seealso{
\link[dsShareServer]{nextDS}, \link[dsShareServer]{isDataEncodedDS}
}
