% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEncryptingDataDS.R
\name{removeEncryptingDataDS}
\alias{removeEncryptingDataDS}
\title{Remove data used to encrypt some parameters}
\usage{
removeEncryptingDataDS(master_mode = TRUE)
}
\arguments{
\item{master_mode}{a boolean argument set to true by default
a two servers}
}
\description{
This server function can be used to remove the data used to encrypt a parameter between
}
